hex-l2: factor out htp_tensor_flush for reuse (if needed)
