+++ ggml/src/ggml-hexagon/htp/CMakeLists.txt
@@ -20,6 +20,7 @@ add_library(${HTP_LIB} SHARED
     work-queue.c
     dma-queue.c
     hmx-queue.c
+    htp-tensor.c
     matmul-ops.c
     flash-attn-ops.c
     gated-delta-net-ops.c
