+++ ggml/src/ggml-hexagon/htp/act-ops.c
@@ -767,8 +767,6 @@ static int execute_op_activations_f32(struct htp_ops_context * octx) {
 }
 
 int op_activations(struct htp_ops_context * octx) {
-    // htp_tensor_make_clean(octx->dst);
-
     switch (octx->src[0]->type) {
         case HTP_TYPE_F32:
             return execute_op_activations_f32(octx);
